# RFC: call plot and render once in module to precompile those routines