@vchuravy 2015-09-01
Maybe instead Base.precompile would be helpful? But I see that is the approach #671 followed and I don't know why that was closed.