@spencerlyon2 2015-09-01
Yeah [I tried](https://github.com/spencerlyon2/Gadfly.jl/tree/sl/breakups) something like that too. 

My strategy in the other branch was two fold:

- break large functions into small chunks
- Precompile those chunks as well all the `render` methods for `Geom`s and `Guide`s

I saw almost no speedup

I think the problem was that there are just so many functions those methods call out too and I didn't catch all of them in my `Base.precompile` net.

So, instead of spending a long time trying to figure out just what methods needed to be precompiled, I let the compiler do it for me! The takeaway I have is that we can trick the (pre)-compiler into making sure certain routines are precompiled by just running the code we want to be fast (in this case the call to `plot` and `render`).