@dcjones 2015-09-07
Merged #673 which supersedes this, but thanks for investigating!