@spencerlyon2 2015-09-08
Awesome, glad @timholy was willing to do by hand what I was too lazy to determine. I just made the compiler find the right functions to pre-compile, @timholy listed out all 700+ of them! :clap: